Added framework for validations and setting/getting values
Items ask their definition for:
- validation of the current value
- where to write the value
- where to read the value

All items have a JSON data property and data["value"] is used as the default storage if no getter or setter is defined on the Definition::Field

+++ app/models/data_structures/item.rb
@@ -3,20 +3,16 @@ require "ancestry"
 module DataStructures
   class Item < ApplicationRecord
     has_ancestry
-    belongs_to :container, polymorphic: true
-    serialize :definition_configuration, type: Hash, coder: JSON
-    serialize :data, type: Hash, coder: JSON
-    has_rich_text :rich_text_value
-    has_one_attached :attachment_value
-    has_many_attached :attachment_values
-    belongs_to :model, polymorphic: true, optional: true
-    validate :container do |value|
-      value.errors.add :container, :is_not_a_container unless value.container.is_a? DataStructures::Container
-    end
-    after_save if: :saved_change_to_definition_configuration? do
-      create_items_for definition if definition.respond_to? :items
-    end
+    def items = children.order(:position)
 
+    belongs_to :container, polymorphic: true
+    validate :container do |item|
+      item.errors.add :container, :is_not_a_container unless item.container.is_a? DataStructures::Container
+    end
+    after_save :create_items_for_definition, if: :saved_change_to_definition_configuration?
+
+    serialize :definition_configuration, type: Hash, coder: JSON
+    attribute :definition
     def definition = @definition ||= DataStructures::Definition.load(definition_configuration)
 
     def definition=(definition)
@@ -24,11 +20,26 @@ module DataStructures
       self.definition_configuration = Definition.dump(definition)
     end
 
-    def items = children.order(:position)
+    attribute :value
+    def value = definition.get_value_for(self)
+
+    def value=(value)
+      definition.set_value_for(self, value)
+    end
+    validate :value do |item|
+      definition&.validate_item(item)
+    end
+
+    serialize :data, type: Hash, coder: JSON
+    has_rich_text :rich_text_value
+    has_one_attached :attachment_value
+    has_many_attached :attachment_values
+    belongs_to :model, polymorphic: true, optional: true
 
     private
 
-    def create_items_for definition
+    def create_items_for_definition
+      return unless definition.respond_to? :items
       items.where(position: definition.items.size..).destroy_all
       definition.items.each_with_index do |item, position|
         items.where(position: position).first_or_initialize.update! container: container, parent: self, definition: item
